change call count product by API, merge APIs called on componentDidMount and handleClick function
