+++ src/containers/ProductCategory.js
@@ -1,6 +1,7 @@
 import React, { Component, Fragment } from 'react';
 import {connect} from 'react-redux';
 import Product from '../components/product';
+import Pagination from '../components/pagination';
 import * as API from '../api';
 import { 
     getCategoriesSuccess,
@@ -26,30 +27,15 @@ class ProductCategory extends Component {
             .then(res => this.props.getCategoriesSuccess(res.data))
             .catch(err => this.props.getCategoriesError(err)) 
 
-        this.props.loading(true);
- 
-        API.get(`http://localhost:3000/categories/${this.props.match.params.slug}/products?_page=${this.state.currentPage}&_limit=${this.state.itemPage}`)
-        .then(res => { 
-            this.props.getProductCategorySuccess(res.data.products);
-            this.props.getAmountCategories(res.data.totalItems); 
-            this.props.loading(false);  
-        })
-        .catch(err => this.props.getProductCategoryError(err)) 
+        this.getProductsPagination(this.props.match.params.slug, this.state.currentPage, this.state.itemPage);
     }
     
     // take event change params of router
     componentWillReceiveProps(nextProps){
         if(nextProps.match.params.slug !== this.props.match.params.slug){
 
-            this.props.loading(true);
-            
-            API.get(`http://localhost:3000/categories/${this.props.match.params.slug}/products?_page=${this.state.currentPage}&_limit=${this.state.itemPage}`)
-            .then(res => { 
-                this.props.getProductCategorySuccess(res.data.products);
-                this.props.getAmountCategories(res.data.totalItems); 
-                this.props.loading(false);  
-            })
-            .catch(err => this.props.getProductCategoryError(err)) 
+            this.getProductsPagination(nextProps.match.params.slug, this.state.currentPage, this.state.itemPage);
+
         }
     }
 
@@ -62,15 +48,19 @@ class ProductCategory extends Component {
         this.setState({
             currentPage: Number(event.target.id)
         },() => {
-            this.props.loading(true);  
-            API.get(`http://localhost:3000/categories/${this.props.match.params.slug}/products?_page=${this.state.currentPage}&_limit=${this.state.itemPage}`)
-            .then(res => { 
-                this.props.getProductCategorySuccess(res.data.products);
-                this.props.getAmountCategories(res.data.totalItems); 
-                this.props.loading(false);  
-            })
-            .catch(err => this.props.getProductCategoryError(err)) 
+            this.getProductsPagination(this.props.match.params.slug, this.state.currentPage, this.state.itemPage);
         });
+    }
+
+    getProductsPagination = (productID, page, limit) => {
+        this.props.loading(true);  
+        API.get(`http://localhost:3000/categories/${productID}/products?_page=${page}&_limit=${limit}`)
+        .then(res => { 
+            this.props.getProductCategorySuccess(res.data.products);
+            this.props.getAmountCategories(res.data.totalItems); 
+            this.props.loading(false);  
+        })
+        .catch(err => this.props.getProductCategoryError(err)) 
     }
 
     render() {
@@ -81,11 +71,6 @@ class ProductCategory extends Component {
             pageNumbers.push(i);
         }
 
-        const renderPageNumbers = pageNumbers.map(number => {
-            return  <li className="page-item" key={number}>
-                        <a className="page-link" href="#" id={number} onClick={this.handleClick}>{number}</a>
-                    </li>
-        })
         return (
             <Fragment>
                 <div className="container">
@@ -102,23 +87,7 @@ class ProductCategory extends Component {
                         }
                         </div>
                     </div>
-                    <nav aria-label="Page navigation example" style={{marginTop: 50}}>
-                        <ul className="pagination" style={{justifyContent: 'center'}}>
-                            <li className="page-item">
-                                <a id='1' onClick={this.handleClick} className="page-link" href="#" aria-label="Previous">
-                                    <span aria-hidden="true">&laquo;</span>
-                                    <span className="sr-only">Previous</span>
-                                </a>
-                            </li>
-                            {renderPageNumbers}
-                            <li className="page-item">
-                                <a id={pageNumbers[pageNumbers.length - 1]} onClick={this.handleClick} className="page-link" href="#" aria-label="Next">
-                                    <span aria-hidden="true">&raquo;</span>
-                                    <span className="sr-only" >Next</span>
-                                </a>
-                            </li>
-                        </ul>
-                    </nav>
+                    <Pagination pageNumbers={pageNumbers}  handleClick={this.handleClick}/>
                 </div>
             </Fragment>
         );
